Update README.md
## Disclaimer

This GitHub repository is provided for educational purposes and as a reference for my coursework at The University of Texas at San Antonio. I am not responsible for how this code is used by others. Plagiarism, copying, or any form of academic dishonesty is strictly discouraged and against the policies of The University of Texas at San Antonio. 

Please use the code in this repository responsibly and ensure that you adhere to your school's code of conduct and academic integrity policies. Any consequences resulting from misuse or plagiarism are the sole responsibility of the individuals involved.